+++ test/hotspot/jtreg/compiler/cha/StrengthReduceInterfaceCall.java
@@ -28,44 +28,33 @@
  *          java.base/jdk.internal.misc
  *          java.base/jdk.internal.vm.annotation
  * @library /test/lib /
+ * @compile Utils.java
  * @build sun.hotspot.WhiteBox
  * @run driver jdk.test.lib.helpers.ClassFileInstaller sun.hotspot.WhiteBox
  *
  * @run main/othervm -Xbootclasspath/a:. -XX:+IgnoreUnrecognizedVMOptions -XX:+UnlockDiagnosticVMOptions
  *                   -XX:+PrintCompilation -XX:+PrintInlining -XX:+TraceDependencies -verbose:class -XX:CompileCommand=quiet
- *                   -XX:CompileCommand=compileonly,*::test -XX:CompileCommand=compileonly,*::m -XX:CompileCommand=dontinline,*::test
- *                   -Xbatch -XX:+WhiteBoxAPI -Xmixed
+ *                   -XX:CompileCommand=compileonly,*::m
+ *                   -XX:CompileCommand=compileonly,*::test -XX:CompileCommand=dontinline,*::test
+ *                   -XX:CompileCommand=compileonly,*::testHelper -XX:CompileCommand=inline,*::testHelper
+ *                   -Xbatch -Xmixed -XX:+WhiteBoxAPI
  *                   -XX:-TieredCompilation
  *                      compiler.cha.StrengthReduceInterfaceCall
  *
  * @run main/othervm -Xbootclasspath/a:. -XX:+IgnoreUnrecognizedVMOptions -XX:+UnlockDiagnosticVMOptions
  *                   -XX:+PrintCompilation -XX:+PrintInlining -XX:+TraceDependencies -verbose:class -XX:CompileCommand=quiet
- *                   -XX:CompileCommand=compileonly,*::test -XX:CompileCommand=compileonly,*::m -XX:CompileCommand=dontinline,*::test
- *                   -Xbatch -XX:+WhiteBoxAPI -Xmixed
+ *                   -XX:CompileCommand=compileonly,*::m
+ *                   -XX:CompileCommand=compileonly,*::test -XX:CompileCommand=dontinline,*::test
+ *                   -XX:CompileCommand=compileonly,*::testHelper -XX:CompileCommand=inline,*::testHelper
+ *                   -Xbatch -Xmixed -XX:+WhiteBoxAPI
  *                   -XX:+TieredCompilation -XX:TieredStopAtLevel=1
  *                      compiler.cha.StrengthReduceInterfaceCall
  */
 package compiler.cha;
 
-import jdk.internal.misc.Unsafe;
-import jdk.internal.org.objectweb.asm.ClassWriter;
-import jdk.internal.org.objectweb.asm.MethodVisitor;
 import jdk.internal.vm.annotation.DontInline;
-import sun.hotspot.WhiteBox;
-import sun.hotspot.code.NMethod;
 
-import java.io.IOException;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
-import java.lang.reflect.Method;
-import java.util.HashMap;
-import java.util.concurrent.Callable;
-
-import static jdk.test.lib.Asserts.*;
-import static jdk.internal.org.objectweb.asm.ClassWriter.*;
-import static jdk.internal.org.objectweb.asm.Opcodes.*;
+import static compiler.cha.Utils.*;
 
 public class StrengthReduceInterfaceCall {
     public static void main(String[] args) {
@@ -76,6 +65,7 @@ public class StrengthReduceInterfaceCall {
         run(ThreeLevelHierarchyAbstractVsDefault.class);
         run(ThreeLevelDefaultHierarchy.class);
         run(ThreeLevelDefaultHierarchy1.class);
+        System.out.println("TEST PASSED");
     }
 
     public static class ObjectToString extends ATest<ObjectToString.I> {
@@ -96,7 +86,7 @@ public class StrengthReduceInterfaceCall {
         static class DJ2 implements J { public String toString() { return "DJ2"; }}
 
         @Override
-        public Object test(I i) { return ObjectToStringHelper.test(i); /* invokeinterface I.toString() */ }
+        public Object test(I i) { return ObjectToStringHelper.testHelper(i); /* invokeinterface I.toString() */ }
 
         @TestCase
         public void testMono() {
@@ -185,7 +175,7 @@ public class StrengthReduceInterfaceCall {
 
         @Override
         public Object test(I i) {
-            return ObjectHashCodeHelper.test(i); /* invokeinterface I.hashCode() */
+            return ObjectHashCodeHelper.testHelper(i); /* invokeinterface I.hashCode() */
         }
 
         @TestCase
@@ -733,293 +723,6 @@ public class StrengthReduceInterfaceCall {
                 test(j);
             });
             assertCompiled();
-        }
-    }
-
-    /* =========================================================== */
-
-    interface Action {
-        int run();
-    }
-
-    public static final Unsafe U = Unsafe.getUnsafe();
-
-    interface Test<T> {
-        void call(T o);
-        T receiver(int id);
-
-        default Runnable monomophic() {
-            return () -> {
-                call(receiver(0)); // 100%
-            };
-        }
-
-        default Runnable bimorphic() {
-            return () -> {
-                call(receiver(0)); // 50%
-                call(receiver(1)); // 50%
-            };
-        }
-
-        default Runnable polymorphic() {
-            return () -> {
-                for (int i = 0; i < 23; i++) {
-                    call(receiver(0)); // 92%
-                }
-                call(receiver(1)); // 4%
-                call(receiver(2)); // 4%
-            };
-        }
-
-        default Runnable megamorphic() {
-            return () -> {
-                call(receiver(0)); // 33%
-                call(receiver(1)); // 33%
-                call(receiver(2)); // 33%
-            };
-        }
-
-        default void initialize(Class<?>... cs) {
-            for (Class<?> c : cs) {
-                U.ensureClassInitialized(c);
-            }
-        }
-
-        default void repeat(int cnt, Runnable r) {
-            for (int i = 0; i < cnt; i++) {
-                r.run();
-            }
-        }
-    }
-
-    public static abstract class ATest<T> implements Test<T> {
-        public static final WhiteBox WB = WhiteBox.getWhiteBox();
-
-        public static final Object CORRECT = new Object();
-        public static final Object WRONG   = new Object();
-
-        final Method TEST;
-        private final Class<T> declared;
-        private final Class<?> receiver;
-
-        private final HashMap<Integer, T> receivers = new HashMap<>();
-
-        public ATest(Class<T> declared, Class<?> receiver) {
-            this.declared = declared;
-            this.receiver = receiver;
-            TEST = compute(() -> this.getClass().getDeclaredMethod("test", declared));
-        }
-
-        @DontInline
-        public abstract Object test(T i);
-
-        public abstract void checkInvalidReceiver();
-
-        public T receiver(int id) {
-            return receivers.computeIfAbsent(id, (i -> {
-                try {
-                    MyClassLoader cl = (MyClassLoader) receiver.getClassLoader();
-                    Class<?> sub = cl.subclass(receiver, i);
-                    return (T)sub.getDeclaredConstructor().newInstance();
-                } catch (Exception e) {
-                    throw new Error(e);
-                }
-            }));
-        }
-
-
-        public void compile(Runnable r) {
-            while (!WB.isMethodCompiled(TEST)) {
-                for (int i = 0; i < 100; i++) {
-                    r.run();
-                }
-            }
-            assertCompiled(); // record nmethod info
-        }
-
-        private NMethod prevNM = null;
-
-        public void assertNotCompiled() {
-            NMethod curNM = NMethod.get(TEST, false);
-            assertTrue(prevNM != null); // was previously compiled
-            assertTrue(curNM == null || prevNM.compile_id != curNM.compile_id); // either no nmethod present or recompiled
-            prevNM = curNM; // update nmethod info
-        }
-
-        public void assertCompiled() {
-            NMethod curNM = NMethod.get(TEST, false);
-            assertTrue(curNM != null); // nmethod is present
-            assertTrue(prevNM == null || prevNM.compile_id == curNM.compile_id); // no recompilations if nmethod present
-            prevNM = curNM; // update nmethod info
-        }
-
-        @Override
-        public void call(T i) {
-            assertTrue(test(i) != WRONG);
-        }
-    }
-
-    @Retention(value = RetentionPolicy.RUNTIME)
-    public @interface TestCase {}
-
-    static void run(Class<?> test) {
-        try {
-            for (Method m : test.getDeclaredMethods()) {
-                if (m.isAnnotationPresent(TestCase.class)) {
-                    System.out.println(m.toString());
-                    ClassLoader cl = new MyClassLoader(test);
-                    Class<?> c = cl.loadClass(test.getName());
-                    c.getMethod(m.getName()).invoke(c.getDeclaredConstructor().newInstance());
-                }
-            }
-        } catch (Exception e) {
-            throw new Error(e);
-        }
-    }
-
-    static class ObjectToStringHelper {
-        static Object test(Object o) {
-            throw new Error("not used");
-        }
-    }
-    static class ObjectHashCodeHelper {
-        static int test(Object o) {
-        throw new Error("not used");
-    }
-    }
-
-    static final class MyClassLoader extends ClassLoader {
-        private final Class<?> test;
-
-        MyClassLoader(Class<?> test) {
-            this.test = test;
-        }
-
-        static String intl(String s) {
-            return s.replace('.', '/');
-        }
-
-        Class<?> subclass(Class<?> c, int id) {
-            String name = c.getName() + id;
-            Class<?> sub = findLoadedClass(name);
-            if (sub == null) {
-                ClassWriter cw = new ClassWriter(COMPUTE_MAXS | COMPUTE_FRAMES);
-                cw.visit(52, ACC_PUBLIC | ACC_SUPER, intl(name), null, intl(c.getName()), null);
-
-                { // Default constructor: <init>()V
-                    MethodVisitor mv = cw.visitMethod(ACC_PUBLIC, "<init>", "()V", null, null);
-                    mv.visitCode();
-                    mv.visitVarInsn(ALOAD, 0);
-                    mv.visitMethodInsn(INVOKESPECIAL, intl(c.getName()), "<init>", "()V", false);
-                    mv.visitInsn(RETURN);
-                    mv.visitMaxs(0, 0);
-                    mv.visitEnd();
-                }
-
-                byte[] classFile = cw.toByteArray();
-                return defineClass(name, classFile, 0, classFile.length);
-            }
-            return sub;
-        }
-
-        protected Class<?> loadClass(String name, boolean resolve)
-                throws ClassNotFoundException
-        {
-            // First, check if the class has already been loaded
-            Class<?> c = findLoadedClass(name);
-            if (c == null) {
-                try {
-                    c = getParent().loadClass(name);
-                    if (name.endsWith("ObjectToStringHelper")) {
-                        ClassWriter cw = new ClassWriter(COMPUTE_MAXS | COMPUTE_FRAMES);
-                        cw.visit(52, ACC_PUBLIC | ACC_SUPER, intl(name), null, "java/lang/Object", null);
-
-                        {
-                            MethodVisitor mv = cw.visitMethod(ACC_PUBLIC | ACC_STATIC, "test", "(Ljava/lang/Object;)Ljava/lang/Object;", null, null);
-                            mv.visitCode();
-                            mv.visitVarInsn(ALOAD, 0);
-                            mv.visitMethodInsn(INVOKEINTERFACE, intl(test.getName()) + "$I", "toString", "()Ljava/lang/String;", true);
-                            mv.visitInsn(ARETURN);
-                            mv.visitMaxs(0, 0);
-                            mv.visitEnd();
-                        }
-
-                        byte[] classFile = cw.toByteArray();
-                        return defineClass(name, classFile, 0, classFile.length);
-                    } else if (name.endsWith("ObjectHashCodeHelper")) {
-                        ClassWriter cw = new ClassWriter(COMPUTE_MAXS | COMPUTE_FRAMES);
-                        cw.visit(52, ACC_PUBLIC | ACC_SUPER, intl(name), null, "java/lang/Object", null);
-
-                        {
-                            MethodVisitor mv = cw.visitMethod(ACC_PUBLIC | ACC_STATIC, "test", "(Ljava/lang/Object;)I", null, null);
-                            mv.visitCode();
-                            mv.visitVarInsn(ALOAD, 0);
-                            mv.visitMethodInsn(INVOKEINTERFACE, intl(test.getName()) + "$I", "hashCode", "()I", true);
-                            mv.visitInsn(IRETURN);
-                            mv.visitMaxs(0, 0);
-                            mv.visitEnd();
-                        }
-
-                        byte[] classFile = cw.toByteArray();
-                        return defineClass(name, classFile, 0, classFile.length);
-                    } else if (c == test || name.startsWith(test.getName())) {
-                        try {
-                            String path = name.replace('.', '/') + ".class";
-                            byte[] classFile = getParent().getResourceAsStream(path).readAllBytes();
-                            return defineClass(name, classFile, 0, classFile.length);
-                        } catch (IOException e) {
-                            throw new Error(e);
-                        }
-                    }
-                } catch (ClassNotFoundException e) {
-                    // ClassNotFoundException thrown if class not found
-                    // from the non-null parent class loader
-                }
-
-                if (c == null) {
-                    // If still not found, then invoke findClass in order
-                    // to find the class.
-                    c = findClass(name);
-                }
-            }
-            if (resolve) {
-                resolveClass(c);
-            }
-            return c;
-        }
-    }
-
-    public interface RunnableWithException {
-        void run() throws Throwable;
-    }
-
-    public static void shouldThrow(Class<? extends Throwable> expectedException, RunnableWithException r) {
-        try {
-            r.run();
-            throw new AssertionError("Exception not thrown: " + expectedException.getName());
-        } catch(Throwable e) {
-            if (expectedException == e.getClass()) {
-                // success: proper exception is thrown
-            } else {
-                throw new Error(expectedException.getName() + " is expected", e);
-            }
-        }
-    }
-
-    public static MethodHandle unsafeCastMH(Class<?> cls) {
-        try {
-            MethodHandle mh = MethodHandles.identity(Object.class);
-            return MethodHandles.explicitCastArguments(mh, mh.type().changeReturnType(cls));
-        } catch (Throwable e) {
-            throw new Error(e);
-        }
-    }
-
-    static <T> T compute(Callable<T> c) {
-        try {
-            return c.call();
-        } catch (Exception e) {
-            throw new Error(e);
         }
     }
 }
